Consolidate dataset bias utilities
