+++ docs/Plan.md
@@ -564,7 +564,7 @@ Combine 1-4 and the *effective* context limit becomes hardware bandwidth, not mo
     Analog device detection caches the discovered list to avoid repeated queries.
 60. **Adversarial robustness suite**: Generate gradient-based adversarial prompts and measure model degradation. Acceptable drop is <5% accuracy on the evaluation harness.
 61. **Bias-aware dataset filtering**: Add `DatasetBiasDetector` to compute representation metrics and filter skewed samples. Goal is <5% disparity across demographic slices after filtering.
-61a. **Dataset bias mitigation**: `DataBiasMitigator` reweights or filters entries based on these scores. `download_triples()` now applies the mitigator before storing new files.
+61a. **Dataset bias mitigation**: `DataBiasMitigator` (now consolidated in `dataset_bias_detector.py`) reweights or filters entries based on these scores. `download_triples()` now applies the mitigator before storing new files.
 61b. **Fairness gap visualizer**: `fairness.FairnessVisualizer` plots demographic parity and opportunity gaps. `dataset_summary.py --fairness-report` saves the charts under `docs/datasets/`; they appear in the lineage and memory dashboards for quick inspection.
 
 61e. **Fairness adaptation pipeline**: `FairnessAdaptationPipeline` streams bias and cognitive load metrics during ingestion and adjusts `ActiveDataSelector` weights. Bias scores are cached and weight updates are vectorised for faster adaptation. `DatasetLineageManager` logs fairness before and after adaptation, showing improved demographic parity on toy datasets.
