+++ src/__init__.py
@@ -207,9 +207,9 @@ from .dataset_bias_detector import (
     text_bias_score,
     file_bias_score,
     DatasetBiasDetector,
+    DataBiasMitigator,
 )
 from .dataset_weight_agent import DatasetWeightAgent
-from .data_bias_mitigator import DataBiasMitigator
 from .data_poison_detector import DataPoisonDetector
 from .auto_labeler import AutoLabeler
 from .graphql_memory_gateway import GraphQLMemoryGateway
